function: complete implementation for ft_strcpy
